apps/backend/pyproject.toml
[project]
name = "compliance-engine"
version = "0.1.0"
description = "AI Compliance Manager Backend"
authors = [{ name = "Arun", email = "arun@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "pydantic-settings>=2.1.0",
    "pydantic[email]>=2.5.3",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "openai>=1.10.0",
    "google-generativeai>=0.3.2",
    "aiosqlite>=0.19.0",
    "web3>=6.15.1",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "numpy>=1.26.3",
    "pandas>=2.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
]

[tool.poetry]
name = "compliance-engine"
version = "0.1.0"
description = "AI Compliance Manager Backend"
authors = ["Arun <arun@example.com>"]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"
